competitive-programming-assistant/requirements.txt
fastmcp
pydantic
python-dotenv
httpx
aiohttp
matplotlib
seaborn
numpy
ics